Enhance: Improve file structure
