+++ client/src/components/Authentication/Signin.js
@@ -1,5 +1,5 @@
 import React, { useState } from 'react';
-import { useAuth } from '../contexts/AuthContext';
+import { useAuth } from '../../contexts/AuthContext';
 import {
   Button,
   FormControl,
@@ -10,22 +10,16 @@ import {
   Stack,
 } from '@chakra-ui/react';
 
-const SignUp = (props) => {
+const Login = (props) => {
   const [email, setEmail] = useState('');
   const [password, setPassword] = useState('');
-  const [name, setName] = useState('');
 
-  const { signup, googleLogin } = useAuth();
+  const { login, googleLogin } = useAuth();
 
-  const handleSignUp = async (event) => {
+  const handleSignIn = async (event) => {
     event.preventDefault();
     try {
-      const userCredential = await signup(email, password);
-      const user = userCredential.user;
-      await user.updateProfile({
-        displayName: name,
-      });
-      window.location.assign('/');
+      await login(email, password);
     } catch (err) {
       alert(err.message);
     }
@@ -41,12 +35,8 @@ const SignUp = (props) => {
 
   return (
     <div>
-      <Heading>Sign-up</Heading>
-      <form
-        onSubmit={async (e) => {
-          await handleSignUp(e);
-        }}
-      >
+      <Heading>Sign-in</Heading>
+      <form onSubmit={(e) => handleSignIn(e)}>
         <FormControl mt={4} isRequired>
           <FormLabel>Email address</FormLabel>
           <Input
@@ -55,15 +45,6 @@ const SignUp = (props) => {
             placeholder="Enter email address"
             value={email}
             onChange={(e) => setEmail(e.target.value)}
-          />
-        </FormControl>
-        <FormControl mt={4} isRequired>
-          <FormLabel>Name</FormLabel>
-          <Input
-            size="md"
-            placeholder="Enter Name"
-            value={name}
-            onChange={(e) => setName(e.target.value)}
           />
         </FormControl>
         <FormControl mt={4} isRequired>
@@ -79,7 +60,7 @@ const SignUp = (props) => {
         <br />
         <Stack>
           <Button type="submit" colorScheme="blue">
-            Sign Up
+            Sign In
           </Button>
           <Button onClick={() => handleGoogleLogin()}>
             <img
@@ -96,11 +77,11 @@ const SignUp = (props) => {
         </Stack>
       </form>
       <br />
-      <Link onClick={() => props.showSignupHandler(false)} color="blue.400">
-        Already have an account? Sign-in
+      <Link onClick={() => props.showSignupHandler(true)} color="blue.400">
+        Don't have an account? Sign-up
       </Link>
     </div>
   );
 };
 
-export default SignUp;
+export default Login;
